Backend/match-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
SQLAlchemy>=1.4,<1.5
databases==0.6.1
psycopg2-binary==2.9.7
pydantic==1.10.9
PyJWT==2.8.0
